Refactor main -> rewards
