+++ src/frontend/index.js
@@ -4,6 +4,6 @@ import 'bootstrap-icons/font/bootstrap-icons.css';
 
 import 'pikaday/css/pikaday.css';
 
-import './src/js/main.js';
+import './src/js/rewards.js';
 // Import our custom CSS
 import './src/scss/styles.scss';
