+++ src/frontend/src/js/rewards.js
